Bypass ban by nominatim.openstreetmap.org on iOS
When Forecast.io is used as Weather provider - nominatim.openstreetmap.org is used to get location name. The service is banned Pebble iOS app due to some past abuse and the call failed causing weather call to fail as well. This change adds custom user-agent header that bypasses the ban